requirements.txt
dj-database-url==0.4.1
Django==1.9.9
gunicorn==19.6.0
virtualenv==15.0.3
whitenoise==3.2.1
psycopg2==2.5.4
djangorestframework==3.4.6
django-cors-headers==1.1.0
django-filter==0.15.2
django-crispy-forms==1.6.0
pytz==2016.7